+++ src/main/java/web/app/onlinebookshop/service/impl/CategoryServiceImpl.java
@@ -53,6 +53,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't delete category with id  " + id);
+        }
         categoryRepository.deleteById(id);
     }
